+++ Frontend/src/components/plan/Calendar.js
@@ -3,7 +3,8 @@ import { DateRange } from 'react-date-range';
 import 'react-date-range/dist/styles.css';
 import 'react-date-range/dist/theme/default.css';
 import axios from 'axios';
-import { Button, List } from 'antd';
+import { Button } from 'antd';
+import DateList from './DateList';
 
 const Calendar = ({ onClose }) => {
   const [selection, setSelection] = useState({
@@ -36,22 +37,6 @@ const Calendar = ({ onClose }) => {
     }
   };
 
-  const start = new Date(selection.startDate);
-  const end = new Date(selection.endDate);
-  const dayCount = Math.floor((end - start) / (1000 * 60 * 60 * 24)) + 1;
-
-  const dailyInfo = [];
-
-  for (let i = 0; i < dayCount; i++) {
-    const currentDate = new Date(start);
-    currentDate.setDate(currentDate.getDate() + i);
-
-    dailyInfo.push({
-      date: currentDate,
-      // date: currentDate.toLocaleDateString(), // 일자를 원하는 형식으로 표시
-    });
-  }
-
   console.log('startDate:', selection.startDate);
   console.log('endDate:', selection.endDate);
 
@@ -64,21 +49,7 @@ const Calendar = ({ onClose }) => {
         ranges={[selection]}
       />
       <Button onClick={handleConfirm}>확인</Button>
-      <div>
-      <h2>날짜별 예산</h2>
-      <List
-        itemLayout="horizontal"
-        dataSource={dailyInfo}
-        renderItem={(item, index) => (
-          <List.Item>
-            <List.Item.Meta
-              title={`날짜: ${item.date.toLocaleDateString()}`}
-            />
-          </List.Item>
-        )}
-      />
-    </div>
-
+      <DateList startDate={selection.startDate} endDate={selection.endDate} />
     </div>
   );
 };
